Adds a reference to the xcode project
Hi,

I have been recently researched the possibility of packing the jxcore-cordova APIs to react-native, just this past days and stumbled upon this. I have to thank you a lot for the work you've done, and wanted to create an issue to add some information about my own findings.

I would like to help as far as I could contribute to your work, using jxcore-chackra to add iOS support, but don't know how much time I can commit.

Thanks a lot and sorry for the little-bit-too-much of a PR to comment this out.